requirements.txt
amqp
asgiref
billiard
boto3
botocore
CacheControl
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
Django
django-celery
django-filter
django-multiselectfield
django-ses
django-storages
djangorestframework
djangorestframework_simplejwt
firebase-admin
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
gunicorn
h11
httpcore
httplib2
idna
jmespath
kombu
msgpack
numpy
packaging
pandas
pillow
prompt_toolkit
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
PyJWT
pyparsing
python-dateutil
python-dotenv
pytz
requests
rsa
s3transfer
six
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
django-admin-logs
django-rangefilter